update as of November 30 2022

+++ application/views/products/admin_dashboard.php
@@ -6,13 +6,15 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <title>Product Dashboard - Admin</title>
         <link rel="stylesheet" href="<?= base_url(); ?>assets/css/home_style.css">
+        <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.1/jquery.min.js"></script>
+        <style>
+            .success{
+                color: green;
+                margin: 20px;
+            }
+        </style>
     </head>
-    <style>
-        .success{
-            color: green;
-            margin: 20px;
-        }
-    </style>
+    
     <body>
         <div class="dashboard-div">
             <h2>Manage Products</h2>
@@ -35,8 +37,8 @@
                     <td><?= $product['count']?></td>
                     <td><?= $product['qty_sold']?></td>
                     <td>
-                        <a href=/products/edit/<?= $product['id']; ?>">edit</a>
-                        <a href="/products/destroy/<?= $product['id']; ?>">remove</a>
+                        <a href="/products/edit/<?= $product['id']; ?>">edit</a>
+                        <a href="/products/destroy/<?= $product['id']; ?>" onclick="return confirm('Are you sure?'); ">remove</a>
                     </td>
                 </tr>
 <?php  } ?>
